First Commit Damage Prevention
I have lost good source in the past due to not backing it up.. I would not recommend using this at all until I have a chance to work with it more. It's missing a lot of implementation but can serve as an example of how things can be done for now.